Layer1/requirements.txt
google-genai
pydantic
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
pytest
